apis fixed wiith error message

+++ controllers/apiControllers.js
@@ -1,9 +1,17 @@
-// var Volcano = require('../database');
-
+var Volcano = require('../database');
+// This should work both there and elsewhere.
+function isEmptyObject(obj) {
+    for (var key in obj) {
+      if (Object.prototype.hasOwnProperty.call(obj, key)) {
+        return false;
+      }
+    }
+    return true;
+  }
 // List of all Volcones
 exports.volcano_list = async function(req, res) {
     try{
-    const sql = "SELECT * FROM  volcanos ORDER BY volcano_name";
+    const sql = "SELECT * FROM  volcanos ORDER BY id";
     Volcano.all(sql, [], (err, rows) => {
         if (err) {
             return console.error(err.message);
@@ -19,14 +27,21 @@ exports.volcano_list = async function(req, res) {
 
 // get a single volcano
 exports.volcano_detail_id = async function(req, res) {
-    try{    
-    const sql = "SELECT * FROM  volcanos WHERE id = ?"; 
-    Volcano.get(sql, [req.params.id], (err, row) => {
-        if (err) {
-            return console.error(err.message);
-        }
-        res.send(row);
-        });
+    try{
+        var query = "select * from volcanos  WHERE id = ?";
+        var k = Volcano.get(query,[req.params.id], (err, row) => {
+            if (err) {
+                return console.error(err.message);
+            }
+
+            console.log(row);
+            if(isEmptyObject(row)){
+                res.status(404);
+                res.send(`{"error": "Volcano ${req.params.id} not found"}`);
+            }
+            else{
+                res.send(row);            }
+            });
     }
     catch(err){
     res.status(500);
@@ -55,13 +70,27 @@ exports.volcano_create = async function(req, res) {
 // Update a volcano by id
 exports.volcano_update_id = async function(req, res) {
     try{
-    const sql = "UPDATE volcanos SET volcano_name = ?, location = ?, history = ?, active = ? WHERE id = ?";
-    Volcano.run(sql, [req.body.volcano_name, req.body.location, req.body.history, req.body.active, req.params.id], function(err) {
-        if (err) {
-            return console.error(err.message);
-        }
-        res.send(`Volcano ${req.params.id} was updated`);
-    });
+        var query = "select * from volcanos  WHERE id = ?";
+        var k = Volcano.get(query,[req.params.id], (err, row) => {
+            if (err) {
+                return console.error(err.message);
+            }
+
+            console.log(row);
+            if(isEmptyObject(row)){
+                res.status(404);
+                res.send(`{"error": "Volcano ${req.params.id} not found"}`);
+            }
+            else{
+                const sql = "UPDATE volcanos SET volcano_name = ?, location = ?, history = ?, active = ? WHERE id = ?";
+                Volcano.run(sql, [req.body.volcano_name, req.body.location, req.body.history, req.body.active, req.params.id], function(err) {
+                    if (err) {
+                        return console.error(err.message);
+                    }
+                    res.send(`Volcano ${req.params.id} was updated`);
+                });
+            }
+            });
     }
     catch(err){
     res.status(500);
@@ -72,13 +101,27 @@ exports.volcano_update_id = async function(req, res) {
 // Delete a volcano by id
 exports.volcano_delete_id = async function(req, res) {
     try{
-    const sql = "DELETE FROM volcanos WHERE id = ?";
-    Volcano.run(sql, [req.params.id], function(err) {
-        if (err) {
-            return console.error(err.message);
-        }
-        res.send(`Volcano ${req.params.id} was deleted`);
-    });
+        var query = "select * from volcanos  WHERE id = ?";
+        var k = Volcano.get(query,[req.params.id], (err, row) => {
+            if (err) {
+                return console.error(err.message);
+            }
+
+            console.log(row);
+            if(isEmptyObject(row)){
+                res.status(404);
+                res.send(`{"error": "Volcano ${req.params.id} not found"}`);
+            }
+            else{
+                const sql = "DELETE FROM volcanos WHERE id = ?";
+                Volcano.run(sql, [req.params.id], function(err) {
+                    if (err) {
+                        return console.error(err.message);
+                    }
+                    res.send(`Volcano ${req.params.id} was deleted`);
+                });
+            }
+            });
     }
     catch(err){
     res.status(500);
